Reinsert dofs/function_map.h includes
