+++ include/deal.II/numerics/error_estimator.h
@@ -22,6 +22,8 @@
 #include <deal.II/base/exceptions.h>
 #include <deal.II/base/function.h>
 
+#include <deal.II/dofs/deprecated_function_map.h>
+
 #include <deal.II/fe/component_mask.h>
 
 #include <map>
