+++ include/deal.II/numerics/matrix_tools.h
@@ -23,6 +23,8 @@
 #include <deal.II/base/function.h>
 #include <deal.II/base/thread_management.h>
 
+#include <deal.II/dofs/deprecated_function_map.h>
+
 #include <deal.II/lac/affine_constraints.h>
 
 #include <map>
